+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Added `dbg`, `list_length` and `working_directory`.
 
 ### Syntax
 
+Fixed parsing of repeated calls, e.g. `f()()` was previously not
+permitted.
+
 Added closures, e.g. `fun(x, y) { x + y; }`.
 
 If the first line of a file starts with `#`, it will be ignored,