+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### Stdlib
 
-Added `dbg`, `list_get`, `list_length` and `working_directory`.
+Added `dbg`, `list_get`, `list_length`, `path_exists` and
+`working_directory`.
 
 ### Syntax